:recycle: BTS-300/refactor: 다이얼로그 FSD 세그먼트 정비
- fsd 원칙에 따라 다이얼로그의 리듀서, 매처 훅, 타입 정의를 components/dialog 슬라이스로 응집시킴
- 공통 상태 로직을 분리하여 여러 피처에서 재사용 가능하도록 개선(작업중)
