+++ src/components/dialog/index.ts
@@ -1,0 +1,4 @@
+export { dialogReducer, initialDialogState } from './model/dialog-reducer';
+export { useDialogMatcher } from './model/use-dialog-matcher';
+export { ConfirmDialog } from './ui/confirm-dialog';
+export type * from './types';
