+++ src/features/qna/components/detail/qna-dropdown.tsx
@@ -2,11 +2,8 @@ import { useReducer } from 'react';
 
 import Image from 'next/image';
 
+import { dialogReducer, initialDialogState } from '@/components/dialog';
 import { DropdownMenu } from '@/components/ui/dropdown-menu';
-import {
-  dialogReducer,
-  initialDialogState,
-} from '@/features/studyrooms/hooks/useDialogReducer';
 
 import { QnAListItem } from '../../types';
